r10484/sofa : Fix a bunch of "loss of data" warnings in modules/
Former-commit-id: 2cc6f27668eeeca977d41ff209897ed1c656d03b

+++ modules/sofa/component/visualmodel/ColorMap.h
@@ -112,7 +112,7 @@ public:
     void prepareLegend();
 
 
-    unsigned int getNbColors() { return entries.size(); }
+    unsigned int getNbColors() { return (unsigned int) entries.size(); }
     Color getColor(unsigned int i) {
         if (i < entries.size()) return entries[i];
         return Color(0.0, 0.0, 0.0, 0.0);
