+++ modules/sofa/component/visualmodel/OglVariable.cpp
@@ -381,7 +381,7 @@ void OglIntVectorVariable::initVisual()
     const helper::vector<GLint>& v = value.getValue();
     const GLint* vptr = v.empty() ? NULL : &(v[0]);
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setIntVector(idShader, idstr.c_str(), v.size(), vptr);
+        (*it)->setIntVector(idShader, idstr.c_str(), (int)v.size(), vptr);
 }
 
 void OglIntVector2Variable::initVisual()
@@ -391,7 +391,7 @@ void OglIntVector2Variable::initVisual()
     const helper::vector<GLint>& v = value.getValue();
     const GLint* vptr = v.empty() ? NULL : &(v[0]);
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setIntVector2(idShader, idstr.c_str(), v.size()/2, vptr);
+        (*it)->setIntVector2(idShader, idstr.c_str(), (int)v.size()/2, vptr);
 }
 
 void OglIntVector3Variable::initVisual()
@@ -401,7 +401,7 @@ void OglIntVector3Variable::initVisual()
     const helper::vector<GLint>& v = value.getValue();
     const GLint* vptr = v.empty() ? NULL : &(v[0]);
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setIntVector3(idShader, idstr.c_str(), v.size()/3, vptr);
+        (*it)->setIntVector3(idShader, idstr.c_str(), (int)v.size()/3, vptr);
 }
 
 void OglIntVector4Variable::initVisual()
@@ -411,7 +411,7 @@ void OglIntVector4Variable::initVisual()
     const helper::vector<GLint>& v = value.getValue();
     const GLint* vptr = v.empty() ? NULL : &(v[0]);
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setIntVector4(idShader, idstr.c_str(), v.size()/4, vptr);
+        (*it)->setIntVector4(idShader, idstr.c_str(), (int)v.size()/4, vptr);
 }
 
 
@@ -488,7 +488,7 @@ void OglFloatVectorVariable::initVisual()
     const helper::vector<float>& v = value.getValue();
     const float* vptr = v.empty() ? NULL : &(v[0]);
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setFloatVector(idShader, idstr.c_str(), v.size(), vptr);
+        (*it)->setFloatVector(idShader, idstr.c_str(), (int)v.size(), vptr);
 }
 
 void OglFloatVector2Variable::initVisual()
@@ -498,7 +498,7 @@ void OglFloatVector2Variable::initVisual()
     const helper::vector<float>& v = value.getValue();
     const float* vptr = v.empty() ? NULL : &(v[0]);
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setFloatVector2(idShader, idstr.c_str(), v.size()/2, vptr);
+        (*it)->setFloatVector2(idShader, idstr.c_str(), (int)v.size()/2, vptr);
 }
 
 void OglFloatVector3Variable::initVisual()
@@ -508,7 +508,7 @@ void OglFloatVector3Variable::initVisual()
     const helper::vector<float>& v = value.getValue();
     const float* vptr = v.empty() ? NULL : &(v[0]);
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setFloatVector3(idShader, idstr.c_str(), v.size()/3, vptr);
+        (*it)->setFloatVector3(idShader, idstr.c_str(), (int)v.size()/3, vptr);
 }
 
 void OglFloatVector4Variable::initVisual()
@@ -518,7 +518,7 @@ void OglFloatVector4Variable::initVisual()
     const helper::vector<float>& v = value.getValue();
     const float* vptr = v.empty() ? NULL : &(v[0]);
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setFloatVector4(idShader, idstr.c_str(), v.size()/4, vptr);
+        (*it)->setFloatVector4(idShader, idstr.c_str(), (int)v.size()/4, vptr);
 }
 
 
@@ -551,7 +551,7 @@ void OglMatrix2Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix2(idShader, idstr.c_str(), v.size()/4, transp, vptr);
+        (*it)->setMatrix2(idShader, idstr.c_str(), (int)v.size()/4, transp, vptr);
 }
 
 OglMatrix3Variable::OglMatrix3Variable()
@@ -582,7 +582,7 @@ void OglMatrix3Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix3(idShader, idstr.c_str(), v.size()/9, transp, vptr);
+        (*it)->setMatrix3(idShader, idstr.c_str(), (int)v.size()/9, transp, vptr);
 }
 
 OglMatrix4Variable::OglMatrix4Variable()
@@ -613,7 +613,7 @@ void OglMatrix4Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix4(idShader, idstr.c_str(), v.size()/16, transp, vptr);
+        (*it)->setMatrix4(idShader, idstr.c_str(), (int)v.size()/16, transp, vptr);
 }
 
 OglMatrix2x3Variable::OglMatrix2x3Variable()
@@ -644,7 +644,7 @@ void OglMatrix2x3Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix2x3(idShader, idstr.c_str(), v.size()/6, transp, vptr);
+        (*it)->setMatrix2x3(idShader, idstr.c_str(), (int)v.size()/6, transp, vptr);
 }
 
 
@@ -676,7 +676,7 @@ void OglMatrix3x2Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix3x2(idShader, idstr.c_str(), v.size()/6, transp, vptr);
+        (*it)->setMatrix3x2(idShader, idstr.c_str(), (int)v.size()/6, transp, vptr);
 }
 
 OglMatrix2x4Variable::OglMatrix2x4Variable()
@@ -707,7 +707,7 @@ void OglMatrix2x4Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix2x4(idShader, idstr.c_str(), v.size()/8, transp, vptr);
+        (*it)->setMatrix2x4(idShader, idstr.c_str(), (int)v.size()/8, transp, vptr);
 }
 
 OglMatrix4x2Variable::OglMatrix4x2Variable()
@@ -738,7 +738,7 @@ void OglMatrix4x2Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix4x2(idShader, idstr.c_str(), v.size()/8, transp, vptr);
+        (*it)->setMatrix4x2(idShader, idstr.c_str(), (int)v.size()/8, transp, vptr);
 }
 
 OglMatrix3x4Variable::OglMatrix3x4Variable()
@@ -769,7 +769,7 @@ void OglMatrix3x4Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix3x4(idShader, idstr.c_str(), v.size()/12, transp, vptr);
+        (*it)->setMatrix3x4(idShader, idstr.c_str(), (int)v.size()/12, transp, vptr);
 }
 
 OglMatrix4x3Variable::OglMatrix4x3Variable()
@@ -800,7 +800,7 @@ void OglMatrix4x3Variable::initVisual()
     const float* vptr = v.empty() ? NULL : &(v[0]);
     const bool transp = transpose.getValue();
     for(std::set<OglShader*>::iterator it = shaders.begin(), iend = shaders.end(); it!=iend; it++)
-        (*it)->setMatrix4x3(idShader, idstr.c_str(), v.size()/12, transp, vptr);
+        (*it)->setMatrix4x3(idShader, idstr.c_str(), (int)v.size()/12, transp, vptr);
 }
 
 } // namespace visual
